pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"



[tool.ruff]
line-length = 120
lint.select = ["E", "F"]
lint.extend-select = ["I"]
lint.ignore = ["E501"]
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 160

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.packages.find]
include = ["focoos**"]


[project]
name = "focoos"
version = "0.14.0"
description = "Focoos SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "Pillow~=10.2.0",
    "supervision~=0.26.0rc4",
    "opencv-python~=4.11.0.86",
    "pydantic~=2.11.3",
    "pydantic-settings~=2.8.1",
    "tqdm~=4.67.1",
    "numpy~=2.2.1",
    "scipy~=1.14.1",
    "psutil~=6.1.1",
    "setuptools~=75.7.0",
    "matplotlib~=3.10.1",
    "colorama~=0.4.6",
    "ipython",
    "fvcore~=0.1.4",
    "pycocotools~=2.0.8",
    "faster_coco_eval~=1.6.5",
    "timm~=0.9.16",
    "tensorboard~=2.19.0",
]

authors = [{ name = "focoos.ai", email = "info@focoos.ai" }]
keywords = [
    "computer_vision",
    "object_detection",
    "segmentation",
    "model_serving",
]

[project.optional-dependencies]
cpu = ["onnxruntime==1.21.0"]
cuda = ["onnxruntime-gpu==1.21.0"]
tensorrt = ["onnxruntime-gpu==1.21.0","tensorrt==10.5.0"]
torch = ["torch==2.4.0","torchvision"]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "ruff",
    "python-dotenv",
    "gradio~=5.25.2",
    "pre-commit~=4.2.0",
    "sniffio~=1.3.1",
    "ipykernel~=6.29.5",
    "tox",
    "tox-uv",
]
docs = [
    "mkdocs>=1.6.0,<2.0.0",
    "mkdocs-material>=9.5.28,<10.0.0",
    "mkdocstrings[python]>=0.29.0,<0.30.0",
    "mkdocs-include-markdown-plugin>=6.2.1,<7.0.0",
]
